# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile/class:
    patch:
      operationId: updatePlayerClass
      tags: [gameState]
      summary: Save the authenticated user's selected class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlayerClassInput"
      responses:
        "204":
          description: Class saved
        "400":
          description: Unsupported class
        "401":
          description: Authentication required
        "503":
          description: Class could not be saved
...
  /performances:
    post:
      operationId: createLyricPerformance
      tags: [performances]
      summary: Generate and store an exact rap take
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLyricPerformanceInput"
      responses:
        "201":
          description: Passing performance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LyricPerformance"
        "400":
          description: Invalid verse or options
        "429":
          description: Three successful takes already used today
        "422":
          description: Verse contains content that cannot be performed
        "409":
          description: Battle result is unavailable or does not match the verse
        "503":
          description: Provider unavailable or exact fidelity failed
  /performances/{performanceId}/share-video/{option}:
    get:
      operationId: getLyricPerformanceShareVideo
      tags: [performances]
      summary: Render a shareable video from a stored passing performance
      parameters:
        - $ref: "#/components/parameters/PerformanceId"
        - $ref: "#/components/parameters/ShareVideoOption"
      responses:
        "200":
          description: Rendered vertical MP4 video
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid performance ID or withScore option
        "404":
          description: Performance not found
        "409":
          description: No stored battle result is linked to this performance
        "503":
          description: Video renderer unavailable
  /performances/{performanceId}/audio:
    get:
      operationId: getLyricPerformanceAudio
      tags: [performances]
      summary: Replay a stored passing performance
      parameters:
        - $ref: "#/components/parameters/PerformanceId"
      responses:
        "200":
          description: Stored MP3 audio
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        "404":
          description: Performance not found
components:
  parameters:
...
    PerformanceId:
      name: performanceId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    ShareVideoOption:
      name: option
      in: path
      required: true
      schema:
        type: string
        enum: [rap, score]
  schemas:
...
    UpdatePlayerClassInput:
      type: object
      required: [class]
      properties:
        class:
          type: string
          enum: [lyrical_assassin]
...
    CreateLyricPerformanceInput:
      type: object
      required: [verse, intro, echoOut]
      properties:
        verse:
          type: string
          minLength: 1
          maxLength: 5000
        intro:
          type: boolean
          default: true
        echoOut:
          type: boolean
          default: true
        battleId:
          type: integer
          minimum: 1
    LyricPerformance:
      type: object
      required: [id, battleId, verse, intro, echoOut, durationMs, audioSizeBytes, audioProvenance, beatManifest, transcript, fidelityPassed, createdAt, remainingToday, audioBase64]
      properties:
        id:
          type: integer
        battleId:
          type: ["integer", "null"]
        verse:
          type: string
        intro:
          type: boolean
        echoOut:
          type: boolean
        durationMs:
          type: integer
        audioSizeBytes:
          type: integer
        audioProvenance:
          $ref: "#/components/schemas/AudioProvenance"
        beatManifest:
          anyOf:
            - $ref: "#/components/schemas/AudioBeatManifest"
            - type: "null"
        transcript:
          type: string
        fidelityPassed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        remainingToday:
          type: integer
        audioBase64:
          type: string
    AudioProvenance:
      type: string
      enum: [original, generated, other, splice, unknown, composite]
    BeatManifestSource:
      type: object
      required: [id, provenance]
      properties:
        id:
          type: string
          minLength: 1
        provenance:
          type: string
          enum: [original, generated, other, splice, unknown]
    AudioBeatManifest:
      type: object
      required: [version, outputId, sources]
      properties:
        version:
          type: integer
          const: 1
        outputId:
          type: string
          minLength: 1
        sources:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/BeatManifestSource"
...